services/agent/requirements.txt
anthropic>=0.25.0
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
web3>=6.0.0
python-dotenv>=1.0.0
